python_analytics/requirements.txt
firebase-admin==6.4.0
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
python-dateutil==2.8.2
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
flask==3.0.0
flask-cors==4.0.0 